AMP-29425 port AMPET-87 to 2.13.1.12

+++ amp/WEB-INF/src/org/digijava/module/aim/services/auditcleaner/AuditCleanerJob.java
@@ -11,22 +11,26 @@ import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
 
 /**
- * 
+ *
  * @author Diego Dimunzio
- * 
+ *
  */
 public class AuditCleanerJob extends ConnectionCleaningJob {
     private static Logger logger = Logger.getLogger(AuditCleanerJob.class);
-    
+
     /**
-     * 
+     *
      */
     @Override
     public void executeInternal(JobExecutionContext contex) throws JobExecutionException {
         String deletetime = FeaturesUtil.getGlobalSettingValue(GlobalSettingsConstants.AUTOMATIC_AUDIT_LOGGER_CLEANUP);
-        logger.info("Starting to delete audit logs older than " + deletetime + " days.............................");
-        AuditLoggerUtil.deleteLogsByPeriod(deletetime);
-        logger.info("Delete Audit logs finished...................................................");
+        if (deletetime.equals("-1")) {
+            logger.info("Delete Audit Logs Option Disabled :: Please select another option");
+        } else {
+            logger.info("Starting to delete audit logs older than " + deletetime + " days...........................");
+            AuditLoggerUtil.deleteLogsByPeriod(deletetime);
+            logger.info("Delete Audit logs finished.................................................");
+        }
     }
 
 }
